+++ src/main/resources/sectionData.sql
@@ -1,5 +1,5 @@
 CREATE TABLE IF NOT EXISTS `sectionData` (
   letter INTEGER,
   sectionIndex INTEGER,
-  data DECIMAL
+  data DOUBLE
 );
